requirements.txt
pandas
requests
PyYAML
scipy
streamlit
